Add compiled data
This is the data generated by the Cleaning_and_compiling.R script. 

It takes the raw benchmarking_data.csv, parses it out into years, pulls all the relevant info, calculates some columns (e.g., participant IDs, consumption and spending indices, etc), and stores it as this data_long.csv file.

This is the key file to be used for analysis in the replication and analysis scripts.